Apply wittra customizations and changes
Mostly things gotten from gleaning at the mender u-boot repo, such as
disabling video and keyboard support entirely, not loading
fdt_addr (also added serial# to list of things that should always be
dynamic. TODO: add ethaddr) from saved environment etc.

Also we try to remove as much of netboot-related things as possible,
as this is not something we are going to be using.